ТTotal refactoring
