+++ pull.sh
@@ -1,0 +1,4 @@
+#docker system prune -a
+#docker-compose stop
+git pull origin main
+#docker-compose up -d
